Project/versio/requirements.txt
argon2-cffi==19.1.0
astroid==2.2.4
bcrypt==3.1.6
cffi==1.12.2
colorama==0.4.1
Django==2.1.7
django-bootstrap4==0.0.8
django-braces==1.13.0
django-ckeditor==5.6.1
django-hitcount==1.3.0
django-js-asset==1.2.2
django-summernote==0.8.11.4
Faker==1.0.4
isort==4.3.12
lazy-object-proxy==1.3.1
mccabe==0.6.1
misaka==2.1.1
Pillow==6.0.0
pycparser==2.19
pylint==2.3.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
typed-ast==1.3.1
wrapt==1.11.1
